+++ src/Bridge/AzureSearch/CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+0.7
+---
+
+ * [BC BREAK] Add support for `ScopingHttpClient` in `SearchStore`
+ * [BC BREAK] The `endpointUrl` parameter for `SearchStore` has been removed
+ * [BC BREAK] The `apiKey` parameter for `SearchStore` has been removed
+ * [BC BREAK] The `apiVersion` parameter for `SearchStore` has been removed
+
 0.4
 ---